requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.1
email-validator==2.2.0
python-dotenv==1.0.1
Werkzeug==3.1.3
WTForms==3.1.2
gunicorn==21.2.0
PyMySQL==1.1.0
Flask-Migrate==4.1.0
requests==2.31.0
mysqlclient==2.2.7
Flask-Mail==0.9.1
